Alert message handler improvised
